refactor(backend): move handlers and database function to connect into their own file
- add dockerfile for backend
